:art: refactor : 패키지의 구성 개선
