+++ README.md
@@ -1,4 +1,4 @@
-# java-racingcar
+# java-racingcar.domain
 
 자동차 경주 미션 저장소
