+++ src/main/java/racingcar/domain/Car.java
@@ -1,6 +1,6 @@
-package racingcar;
+package racingcar.domain;
 
-import validator.CarNameValidator;
+import racingcar.validator.CarNameValidator;
 
 public class Car {
     private static final int START_POSITION = 0;
