+++ src/main/java/racingcar/domain/RandomBasedStrategy.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class RandomBasedStrategy implements MoveStrategy {
