+++ src/main/java/racingcar/validator/CarNameValidator.java
@@ -1,4 +1,4 @@
-package validator;
+package racingcar.validator;
 
 import java.util.Objects;
