+++ src/main/java/racingcar/view/OutputView.java
@@ -1,6 +1,6 @@
-package view;
+package racingcar.view;
 
-import racingcar.Car;
+import racingcar.domain.Car;
 
 import java.util.List;
