+++ src/test/java/racingcar/validator/CarNameValidatorTest.java
@@ -1,4 +1,4 @@
-package validator;
+package racingcar.validator;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
